Remove unneeded MessageListenerThread class
